Fixed bug #1363089 "warnings format '%d' expects type 'int' not 'long unsigned int'"

+++ src/ui/mainscr.cpp
@@ -1298,7 +1298,7 @@ void CInfoPanel::Draw(void)
 			if (NumSelected > (int)UI.SelectedButtons.size()) {
 				char buf[5];
 
-				sprintf(buf, "+%ud", NumSelected - UI.SelectedButtons.size());
+				sprintf(buf, "+%ud", static_cast<unsigned int> (NumSelected - UI.SelectedButtons.size()));
 				VideoDrawText(UI.MaxSelectedTextX, UI.MaxSelectedTextY,
 					UI.MaxSelectedFont, buf);
 			}
